requirements.txt
azure-storage-blob
boto3
